test(jx promote): test helper refactor
