+++ pkg/jx/cmd/controller_workflow_integration_test.go
@@ -3,16 +3,8 @@
 package cmd_test
 
 import (
-	"fmt"
-	"strconv"
-	"strings"
-	"testing"
-	"time"
-
-	"github.com/ghodss/yaml"
 	"github.com/jenkins-x/jx/pkg/apis/jenkins.io/v1"
 	"github.com/jenkins-x/jx/pkg/client/clientset/versioned"
-	typev1 "github.com/jenkins-x/jx/pkg/client/clientset/versioned/typed/jenkins.io/v1"
 	"github.com/jenkins-x/jx/pkg/gits"
 	"github.com/jenkins-x/jx/pkg/helm"
 	"github.com/jenkins-x/jx/pkg/jx/cmd"
@@ -20,9 +12,8 @@ import (
 	"github.com/jenkins-x/jx/pkg/log"
 	"github.com/jenkins-x/jx/pkg/workflow"
 	"github.com/stretchr/testify/assert"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/util/uuid"
+	"testing"
 )
 
 func TestSequentialWorkflow(t *testing.T) {
@@ -39,7 +30,7 @@ func TestSequentialWorkflow(t *testing.T) {
 
 	o := &cmd.ControllerWorkflowOptions{
 		NoWatch:          true,
-		FakePullRequests: NewCreateEnvPullRequestFn(fakeGitProvider),
+		FakePullRequests: cmd.NewCreateEnvPullRequestFn(fakeGitProvider),
 		FakeGitProvider:  fakeGitProvider,
 	}
 
@@ -80,15 +71,15 @@ func TestSequentialWorkflow(t *testing.T) {
 			spec := workflow.Spec
 			assert.Equal(t, 2, len(spec.Steps), "number of steps")
 			if len(spec.Steps) > 0 {
-				assertPromoteStep(t, &spec.Steps[0], "staging")
+				cmd.AssertPromoteStep(t, &spec.Steps[0], "staging")
 			}
 			if len(spec.Steps) > 1 {
-				assertPromoteStep(t, &spec.Steps[1], "production")
+				cmd.AssertPromoteStep(t, &spec.Steps[1], "production")
 			}
 		}
 	}
 
-	a, err := createTestPipelineActivity(jxClient, ns, testOrgName, testRepoName, "master", "1", myFlowName)
+	a, err := cmd.CreateTestPipelineActivity(jxClient, ns, testOrgName, testRepoName, "master", "1", myFlowName)
 	assert.NoError(t, err)
 	if err != nil {
 		return
@@ -99,50 +90,50 @@ func TestSequentialWorkflow(t *testing.T) {
 		return
 	}
 	activities := jxClient.JenkinsV1().PipelineActivities(ns)
-	assertHasPullRequestForEnv(t, activities, a.Name, "staging")
-	assertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
+	cmd.AssertHasPullRequestForEnv(t, activities, a.Name, "staging")
+	cmd.AssertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
 
 	// lets make sure we don't create a PR for production as we have not completed the staging PR yet
 	err = o.Run()
-	assertHasNoPullRequestForEnv(t, activities, a.Name, "production")
+	cmd.AssertHasNoPullRequestForEnv(t, activities, a.Name, "production")
 
 	// still no PR merged so cannot create a PR for production
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
-	assertHasNoPullRequestForEnv(t, activities, a.Name, "production")
+	cmd.AssertHasNoPullRequestForEnv(t, activities, a.Name, "production")
 
 	// test no PR on production until staging completed
-	if !assertSetPullRequestMerged(t, fakeGitProvider, stagingRepo, 1) {
+	if !cmd.AssertSetPullRequestMerged(t, fakeGitProvider, stagingRepo, 1) {
 		return
 	}
 
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
-	assertHasNoPullRequestForEnv(t, activities, a.Name, "production")
+	cmd.AssertHasNoPullRequestForEnv(t, activities, a.Name, "production")
 
-	if !assertSetPullRequestComplete(t, fakeGitProvider, stagingRepo, 1) {
+	if !cmd.AssertSetPullRequestComplete(t, fakeGitProvider, stagingRepo, 1) {
 		return
 	}
 
 	// now lets poll again due to change to the activity to detect the staging is complete
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
 
-	assertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
-	assertHasPullRequestForEnv(t, activities, a.Name, "production")
-	assertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
-	assertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPullRequestForEnv(t, activities, a.Name, "production")
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
+	cmd.AssertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
 
-	if !assertSetPullRequestMerged(t, fakeGitProvider, prodRepo, 1) {
+	if !cmd.AssertSetPullRequestMerged(t, fakeGitProvider, prodRepo, 1) {
 		return
 	}
-	if !assertSetPullRequestComplete(t, fakeGitProvider, prodRepo, 1) {
+	if !cmd.AssertSetPullRequestComplete(t, fakeGitProvider, prodRepo, 1) {
 		return
 	}
 
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
 
-	assertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
-	assertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeSucceeded)
 
-	assertAllPromoteStepsSuccessful(t, activities, a.Name)
+	cmd.AssertAllPromoteStepsSuccessful(t, activities, a.Name)
 }
 
 func TestWorkflowManualPromote(t *testing.T) {
@@ -159,7 +150,7 @@ func TestWorkflowManualPromote(t *testing.T) {
 
 	o := &cmd.ControllerWorkflowOptions{
 		NoWatch:          true,
-		FakePullRequests: NewCreateEnvPullRequestFn(fakeGitProvider),
+		FakePullRequests: cmd.NewCreateEnvPullRequestFn(fakeGitProvider),
 		FakeGitProvider:  fakeGitProvider,
 	}
 
@@ -185,7 +176,7 @@ func TestWorkflowManualPromote(t *testing.T) {
 	jxClient, ns, err := o.JXClientAndDevNamespace()
 	assert.NoError(t, err)
 
-	a, err := createTestPipelineActivity(jxClient, ns, testOrgName, testRepoName, "master", "1", workflowName)
+	a, err := cmd.CreateTestPipelineActivity(jxClient, ns, testOrgName, testRepoName, "master", "1", workflowName)
 	assert.NoError(t, err)
 	if err != nil {
 		return
@@ -196,27 +187,27 @@ func TestWorkflowManualPromote(t *testing.T) {
 		return
 	}
 	activities := jxClient.JenkinsV1().PipelineActivities(ns)
-	assertHasPullRequestForEnv(t, activities, a.Name, "staging")
-	assertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
+	cmd.AssertHasPullRequestForEnv(t, activities, a.Name, "staging")
+	cmd.AssertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
 
 	// lets make sure we don't create a PR for production as its manual
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
-	assertHasNoPullRequestForEnv(t, activities, a.Name, "production")
+	cmd.AssertHasNoPullRequestForEnv(t, activities, a.Name, "production")
 
-	if !assertSetPullRequestMerged(t, fakeGitProvider, stagingRepo, 1) {
+	if !cmd.AssertSetPullRequestMerged(t, fakeGitProvider, stagingRepo, 1) {
 		return
 	}
-	if !assertSetPullRequestComplete(t, fakeGitProvider, stagingRepo, 1) {
+	if !cmd.AssertSetPullRequestComplete(t, fakeGitProvider, stagingRepo, 1) {
 		return
 	}
 
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
 
-	assertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
 
-	assertHasNoPullRequestForEnv(t, activities, a.Name, "production")
-	assertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
-	assertAllPromoteStepsSuccessful(t, activities, a.Name)
+	cmd.AssertHasNoPullRequestForEnv(t, activities, a.Name, "production")
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertAllPromoteStepsSuccessful(t, activities, a.Name)
 
 	// now lets do a manual promotion
 	version := a.Spec.Version
@@ -241,62 +232,49 @@ func TestWorkflowManualPromote(t *testing.T) {
 		return
 	}
 
-	assertHasPullRequestForEnv(t, activities, a.Name, "production")
-	assertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
-	assertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
+	cmd.AssertHasPullRequestForEnv(t, activities, a.Name, "production")
+	cmd.AssertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
+	cmd.AssertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
 
-	assertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
-	assertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
-
-	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
-	assertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
-	assertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
-	assertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
-	assertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
 
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
-	assertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
-	assertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
-	assertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
-	assertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
+	cmd.AssertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
 
-	if !assertSetPullRequestMerged(t, fakeGitProvider, prodRepo, 1) {
+	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
+	cmd.AssertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
+
+	if !cmd.AssertSetPullRequestMerged(t, fakeGitProvider, prodRepo, 1) {
 		return
 	}
 
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
-	assertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
-	assertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
-	assertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
-	assertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
+	cmd.AssertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
 
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
-	assertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
-	assertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
-	assertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
-	assertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
+	cmd.AssertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeSucceeded)
 
-	if !assertSetPullRequestComplete(t, fakeGitProvider, prodRepo, 1) {
+	if !cmd.AssertSetPullRequestComplete(t, fakeGitProvider, prodRepo, 1) {
 		return
 	}
 
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
-	assertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
-	assertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeSucceeded)
-	assertAllPromoteStepsSuccessful(t, activities, a.Name)
-}
-
-func dumpActivity(t *testing.T, activities typev1.PipelineActivityInterface, name string) *v1.PipelineActivity {
-	activity, err := activities.Get(name, metav1.GetOptions{})
-	assert.NoError(t, err)
-	if err != nil {
-		return nil
-	}
-	assert.NotNil(t, activity, "No PipelineActivity found for name %s", name)
-	if activity != nil {
-		dumpFailedActivity(activity)
-	}
-	return activity
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertAllPromoteStepsSuccessful(t, activities, a.Name)
 }
 
 // TestParallelWorkflow lets test promoting to A + B then when A + B is complete then C
@@ -321,7 +299,7 @@ func TestParallelWorkflow(t *testing.T) {
 
 	o := &cmd.ControllerWorkflowOptions{
 		NoWatch:          true,
-		FakePullRequests: NewCreateEnvPullRequestFn(fakeGitProvider),
+		FakePullRequests: cmd.NewCreateEnvPullRequestFn(fakeGitProvider),
 		FakeGitProvider:  fakeGitProvider,
 	}
 
@@ -364,18 +342,18 @@ func TestParallelWorkflow(t *testing.T) {
 			spec := workflow.Spec
 			assert.Equal(t, 3, len(spec.Steps), "number of steps")
 			if len(spec.Steps) > 0 {
-				assertPromoteStep(t, &spec.Steps[0], envNameA)
+				cmd.AssertPromoteStep(t, &spec.Steps[0], envNameA)
 			}
 			if len(spec.Steps) > 1 {
-				assertPromoteStep(t, &spec.Steps[1], envNameB)
+				cmd.AssertPromoteStep(t, &spec.Steps[1], envNameB)
 			}
 			if len(spec.Steps) > 2 {
-				assertPromoteStep(t, &spec.Steps[2], envNameC)
+				cmd.AssertPromoteStep(t, &spec.Steps[2], envNameC)
 			}
 		}
 	}
 
-	a, err := createTestPipelineActivity(jxClient, ns, testOrgName, testRepoName, "master", "1", myFlowName)
+	a, err := cmd.CreateTestPipelineActivity(jxClient, ns, testOrgName, testRepoName, "master", "1", myFlowName)
 	assert.NoError(t, err)
 	if err != nil {
 		return
@@ -386,68 +364,68 @@ func TestParallelWorkflow(t *testing.T) {
 		return
 	}
 	activities := jxClient.JenkinsV1().PipelineActivities(ns)
-	assertHasPullRequestForEnv(t, activities, a.Name, envNameA)
-	assertHasPullRequestForEnv(t, activities, a.Name, envNameB)
-	assertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
+	cmd.AssertHasPullRequestForEnv(t, activities, a.Name, envNameA)
+	cmd.AssertHasPullRequestForEnv(t, activities, a.Name, envNameB)
+	cmd.AssertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
 
 	// lets make sure we don't create a PR for production as we have not completed the staging PR yet
 	err = o.Run()
-	assertHasNoPullRequestForEnv(t, activities, a.Name, envNameC)
+	cmd.AssertHasNoPullRequestForEnv(t, activities, a.Name, envNameC)
 
 	// still no PR merged so cannot create a PR for C until A and B complete
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
-	assertHasNoPullRequestForEnv(t, activities, a.Name, envNameC)
+	cmd.AssertHasNoPullRequestForEnv(t, activities, a.Name, envNameC)
 
 	// test no PR on production until staging completed
-	if !assertSetPullRequestMerged(t, fakeGitProvider, repoA, 1) {
+	if !cmd.AssertSetPullRequestMerged(t, fakeGitProvider, repoA, 1) {
 		return
 	}
 
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
-	assertHasNoPullRequestForEnv(t, activities, a.Name, envNameC)
+	cmd.AssertHasNoPullRequestForEnv(t, activities, a.Name, envNameC)
 
-	if !assertSetPullRequestComplete(t, fakeGitProvider, repoA, 1) {
+	if !cmd.AssertSetPullRequestComplete(t, fakeGitProvider, repoA, 1) {
 		return
 	}
 
 	// now lets poll again due to change to the activity to detect the staging is complete
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
 
-	assertHasNoPullRequestForEnv(t, activities, a.Name, envNameC)
-	assertHasPromoteStatus(t, activities, a.Name, envNameA, v1.ActivityStatusTypeSucceeded)
-	assertHasPromoteStatus(t, activities, a.Name, envNameB, v1.ActivityStatusTypeRunning)
-	assertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
+	cmd.AssertHasNoPullRequestForEnv(t, activities, a.Name, envNameC)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, envNameA, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, envNameB, v1.ActivityStatusTypeRunning)
+	cmd.AssertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
 
-	if !assertSetPullRequestMerged(t, fakeGitProvider, repoB, 1) {
+	if !cmd.AssertSetPullRequestMerged(t, fakeGitProvider, repoB, 1) {
 		return
 	}
-	if !assertSetPullRequestComplete(t, fakeGitProvider, repoB, 1) {
+	if !cmd.AssertSetPullRequestComplete(t, fakeGitProvider, repoB, 1) {
 		return
 	}
 
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
 
 	// C should have started now
-	assertHasPullRequestForEnv(t, activities, a.Name, envNameC)
-	assertHasPromoteStatus(t, activities, a.Name, envNameA, v1.ActivityStatusTypeSucceeded)
-	assertHasPromoteStatus(t, activities, a.Name, envNameB, v1.ActivityStatusTypeSucceeded)
-	assertHasPromoteStatus(t, activities, a.Name, envNameC, v1.ActivityStatusTypeRunning)
+	cmd.AssertHasPullRequestForEnv(t, activities, a.Name, envNameC)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, envNameA, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, envNameB, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, envNameC, v1.ActivityStatusTypeRunning)
 
-	if !assertSetPullRequestMerged(t, fakeGitProvider, repoC, 1) {
+	if !cmd.AssertSetPullRequestMerged(t, fakeGitProvider, repoC, 1) {
 		return
 	}
-	if !assertSetPullRequestComplete(t, fakeGitProvider, repoC, 1) {
+	if !cmd.AssertSetPullRequestComplete(t, fakeGitProvider, repoC, 1) {
 		return
 	}
 
 	// should be complete now
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
 
-	assertHasPromoteStatus(t, activities, a.Name, envNameA, v1.ActivityStatusTypeSucceeded)
-	assertHasPromoteStatus(t, activities, a.Name, envNameB, v1.ActivityStatusTypeSucceeded)
-	assertHasPromoteStatus(t, activities, a.Name, envNameC, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, envNameA, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, envNameB, v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, envNameC, v1.ActivityStatusTypeSucceeded)
 
-	assertAllPromoteStepsSuccessful(t, activities, a.Name)
+	cmd.AssertAllPromoteStepsSuccessful(t, activities, a.Name)
 }
 
 // TestNewVersionWhileExistingWorkflow lets test that we create a new workflow and terminate
@@ -466,7 +444,7 @@ func TestNewVersionWhileExistingWorkflow(t *testing.T) {
 
 	o := &cmd.ControllerWorkflowOptions{
 		NoWatch:          true,
-		FakePullRequests: NewCreateEnvPullRequestFn(fakeGitProvider),
+		FakePullRequests: cmd.NewCreateEnvPullRequestFn(fakeGitProvider),
 		FakeGitProvider:  fakeGitProvider,
 	}
 
@@ -507,15 +485,15 @@ func TestNewVersionWhileExistingWorkflow(t *testing.T) {
 			spec := workflow.Spec
 			assert.Equal(t, 2, len(spec.Steps), "number of steps")
 			if len(spec.Steps) > 0 {
-				assertPromoteStep(t, &spec.Steps[0], "staging")
+				cmd.AssertPromoteStep(t, &spec.Steps[0], "staging")
 			}
 			if len(spec.Steps) > 1 {
-				assertPromoteStep(t, &spec.Steps[1], "production")
+				cmd.AssertPromoteStep(t, &spec.Steps[1], "production")
 			}
 		}
 	}
 
-	a, err := createTestPipelineActivity(jxClient, ns, testOrgName, testRepoName, "master", "1", myFlowName)
+	a, err := cmd.CreateTestPipelineActivity(jxClient, ns, testOrgName, testRepoName, "master", "1", myFlowName)
 	assert.NoError(t, err)
 	if err != nil {
 		return
@@ -526,61 +504,61 @@ func TestNewVersionWhileExistingWorkflow(t *testing.T) {
 		return
 	}
 	activities := jxClient.JenkinsV1().PipelineActivities(ns)
-	assertHasPullRequestForEnv(t, activities, a.Name, "staging")
-	assertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
+	cmd.AssertHasPullRequestForEnv(t, activities, a.Name, "staging")
+	cmd.AssertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
 
 	// lets trigger a new pipeline release which should close the old version
 	aOld := a
-	a, err = createTestPipelineActivity(jxClient, ns, testOrgName, testRepoName, "master", "2", myFlowName)
+	a, err = cmd.CreateTestPipelineActivity(jxClient, ns, testOrgName, testRepoName, "master", "2", myFlowName)
 
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
 
-	assertHasPullRequestForEnv(t, activities, a.Name, "staging")
-	assertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
+	cmd.AssertHasPullRequestForEnv(t, activities, a.Name, "staging")
+	cmd.AssertWorkflowStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
 
 	// lets make sure we don't create a PR for production as we have not completed the staging PR yet
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
-	assertHasNoPullRequestForEnv(t, activities, a.Name, "production")
+	cmd.AssertHasNoPullRequestForEnv(t, activities, a.Name, "production")
 
-	assertWorkflowStatus(t, activities, aOld.Name, v1.ActivityStatusTypeAborted)
+	cmd.AssertWorkflowStatus(t, activities, aOld.Name, v1.ActivityStatusTypeAborted)
 
 	// still no PR merged so cannot create a PR for production
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
-	assertHasNoPullRequestForEnv(t, activities, a.Name, "production")
+	cmd.AssertHasNoPullRequestForEnv(t, activities, a.Name, "production")
 
 	// test no PR on production until staging completed
-	if !assertSetPullRequestMerged(t, fakeGitProvider, stagingRepo, 2) {
+	if !cmd.AssertSetPullRequestMerged(t, fakeGitProvider, stagingRepo, 2) {
 		return
 	}
 
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
-	assertHasNoPullRequestForEnv(t, activities, a.Name, "production")
+	cmd.AssertHasNoPullRequestForEnv(t, activities, a.Name, "production")
 
-	if !assertSetPullRequestComplete(t, fakeGitProvider, stagingRepo, 2) {
+	if !cmd.AssertSetPullRequestComplete(t, fakeGitProvider, stagingRepo, 2) {
 		return
 	}
 
 	// now lets poll again due to change to the activity to detect the staging is complete
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
 
-	assertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
-	assertHasPullRequestForEnv(t, activities, a.Name, "production")
-	assertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
-	assertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPullRequestForEnv(t, activities, a.Name, "production")
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeRunning)
+	cmd.AssertHasPipelineStatus(t, activities, a.Name, v1.ActivityStatusTypeRunning)
 
-	if !assertSetPullRequestMerged(t, fakeGitProvider, prodRepo, 1) {
+	if !cmd.AssertSetPullRequestMerged(t, fakeGitProvider, prodRepo, 1) {
 		return
 	}
-	if !assertSetPullRequestComplete(t, fakeGitProvider, prodRepo, 1) {
+	if !cmd.AssertSetPullRequestComplete(t, fakeGitProvider, prodRepo, 1) {
 		return
 	}
 
 	pollGitStatusAndReactToPipelineChanges(t, o, jxClient, ns)
 
-	assertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
-	assertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "staging", v1.ActivityStatusTypeSucceeded)
+	cmd.AssertHasPromoteStatus(t, activities, a.Name, "production", v1.ActivityStatusTypeSucceeded)
 
-	assertAllPromoteStepsSuccessful(t, activities, a.Name)
+	cmd.AssertAllPromoteStepsSuccessful(t, activities, a.Name)
 }
 
 func TestPullRequestNumber(t *testing.T) {
@@ -603,307 +581,9 @@ func TestPullRequestNumber(t *testing.T) {
 	}
 }
 
-func dumpPipelineMap(o *cmd.ControllerWorkflowOptions) {
-	log.Infof("Dumping PipelineMap {\n")
-	for k, v := range o.PipelineMap() {
-		log.Infof("    Pipeline %s %s\n", k, v.Name)
-	}
-	log.Infof("}\n")
-}
-
 func pollGitStatusAndReactToPipelineChanges(t *testing.T, o *cmd.ControllerWorkflowOptions, jxClient versioned.Interface, ns string) error {
 	o.ReloadAndPollGitPipelineStatuses(jxClient, ns)
 	err := o.Run()
 	assert.NoError(t, err, "Failed to react to PipelineActivity changes")
 	return err
-}
-
-func assertSetPullRequestMerged(t *testing.T, provider *gits.FakeProvider, repository *gits.FakeRepository, prNumber int) bool {
-	fakePR := repository.PullRequests[prNumber]
-	if !assert.NotNil(t, fakePR, "No PullRequest found on repository %s for number #%d", repository.String(), prNumber) {
-		return false
-	}
-	commitLen := len(fakePR.Commits)
-	if !assert.True(t, commitLen > 0, "PullRequest #%d on repository %s has no commits", prNumber, repository.String()) {
-		return false
-	}
-	lastFakeCommit := fakePR.Commits[commitLen-1].Commit
-	if !assert.NotNil(t, lastFakeCommit, "PullRequest #%d on repository %s last commit status has no commits", prNumber, repository.String()) {
-		return false
-	}
-	sha := lastFakeCommit.SHA
-	merged := true
-	fakePR.PullRequest.MergeCommitSHA = &sha
-	fakePR.PullRequest.Merged = &merged
-
-	log.Infof("PR %s is now merged\n", fakePR.PullRequest.URL)
-
-	// validate the fake Git provider concurs
-	testGitInfo := &gits.GitRepositoryInfo{
-		Organisation: repository.Owner,
-		Name:         repository.Name(),
-	}
-	pr, err := provider.GetPullRequest(repository.Owner, testGitInfo, prNumber)
-	assert.NoError(t, err, "Finding PullRequest %d", prNumber)
-	return assert.True(t, pr.Merged != nil && *pr.Merged, "Fake PR %d is merged", prNumber)
-}
-
-func assertSetPullRequestComplete(t *testing.T, provider *gits.FakeProvider, repository *gits.FakeRepository, prNumber int) bool {
-	fakePR := repository.PullRequests[prNumber]
-	if !assert.NotNil(t, fakePR, "No PullRequest found on repository %s for number #%d", repository.String(), prNumber) {
-		return false
-	}
-
-	l := len(fakePR.Commits)
-	if l > 0 {
-		fakePR.Commits[l-1].Status = gits.CommitSatusSuccess
-
-		// ensure the commit is on the repo r
-		lastCommit := fakePR.Commits[l-1]
-		if len(repository.Commits) == 0 {
-			repository.Commits = append(repository.Commits, lastCommit)
-		} else {
-			repository.Commits[len(repository.Commits)-1] = lastCommit
-		}
-		log.Infof("PR %s has commit status success\n", fakePR.PullRequest.URL)
-	}
-
-	// validate the fake Git provider concurs
-	repoOwner := repository.Owner
-	repoName := repository.Name()
-	testGitInfo := &gits.GitRepositoryInfo{
-		Organisation: repoOwner,
-		Name:         repoName,
-	}
-	pr, err := provider.GetPullRequest(repoOwner, testGitInfo, prNumber)
-	assert.NoError(t, err, "Finding PullRequest %d", prNumber)
-	if !assert.NotNil(t, pr, "Could not find PR %d", prNumber) {
-		return false
-	}
-	if !assert.NotNil(t, pr.MergeCommitSHA, "PR %d has no MergeCommitSHA", prNumber) {
-		return false
-	}
-
-	statuses, err := provider.ListCommitStatus(repoOwner, repoName, *pr.MergeCommitSHA)
-	assert.NoError(t, err, "Finding PullRequest %d commit status", prNumber)
-	if assert.True(t, len(statuses) > 0, "PullRequest %d statuses are empty", prNumber) {
-		lastStatus := statuses[len(statuses)-1]
-		return assert.Equal(t, "success", lastStatus.State, "Last commit status of PullRequest 1 at %s", pr.URL)
-	}
-	return false
-}
-
-func setPullRequestClosed(pr *gits.FakePullRequest) {
-	now := time.Now()
-	pr.PullRequest.ClosedAt = &now
-
-	log.Infof("PR %s is now closed\n", pr.PullRequest.URL)
-}
-
-func assertHasPullRequestForEnv(t *testing.T, activities typev1.PipelineActivityInterface, name string, envName string) {
-	activity, err := activities.Get(name, metav1.GetOptions{})
-	if err != nil {
-		assert.NoError(t, err, "Could not find PipelineActivity %s", name)
-		return
-	}
-	for _, step := range activity.Spec.Steps {
-		promote := step.Promote
-		if promote != nil {
-			if promote.Environment == envName {
-				failed := false
-				pullRequestStep := promote.PullRequest
-				if pullRequestStep == nil {
-					assert.Fail(t, "No PullRequest object on PipelineActivity %s for Promote step for Environment %s", name, envName)
-					failed = true
-				}
-				u := pullRequestStep.PullRequestURL
-				log.Infof("Found Promote PullRequest %s on PipelineActivity %s for Environment %s\n", u, name, envName)
-
-				if !assert.True(t, u != "", "No PullRequest URL on PipelineActivity %s for Promote step for Environment %s", name, envName) {
-					failed = true
-				}
-				if failed {
-					dumpFailedActivity(activity)
-				}
-				return
-			}
-		}
-	}
-	assert.Fail(t, "Missing Promote", "No Promote found on PipelineActivity %s for Environment %s", name, envName)
-	dumpFailedActivity(activity)
-}
-
-func assertHasPromoteStatus(t *testing.T, activities typev1.PipelineActivityInterface, name string, envName string, status v1.ActivityStatusType) {
-	activity, err := activities.Get(name, metav1.GetOptions{})
-	if err != nil {
-		assert.NoError(t, err, "Could not find PipelineActivity %s", name)
-		return
-	}
-	for _, step := range activity.Spec.Steps {
-		promote := step.Promote
-		if promote != nil {
-			if promote.Environment == envName {
-				if !assert.Equal(t, string(status), string(promote.Status), "activity status for %s promote %s", name, envName) {
-					dumpFailedActivity(activity)
-				}
-				return
-			}
-		}
-	}
-	assert.Fail(t, "Missing Promote", "No Promote found on PipelineActivity %s for Environment %s", name, envName)
-	dumpFailedActivity(activity)
-}
-
-func assertHasPipelineStatus(t *testing.T, activities typev1.PipelineActivityInterface, name string, status v1.ActivityStatusType) {
-	activity, err := activities.Get(name, metav1.GetOptions{})
-	if err != nil {
-		assert.NoError(t, err, "Could not find PipelineActivity %s", name)
-		return
-	}
-	if !assert.Equal(t, string(status), string(activity.Spec.Status), "activity status for PipelineActivity %s", name) {
-		dumpFailedActivity(activity)
-	}
-}
-
-func assertAllPromoteStepsSuccessful(t *testing.T, activities typev1.PipelineActivityInterface, name string) {
-	activity, err := activities.Get(name, metav1.GetOptions{})
-	if err != nil {
-		assert.NoError(t, err, "Could not find PipelineActivity %s", name)
-		return
-	}
-	assert.Equal(t, string(v1.ActivityStatusTypeSucceeded), string(activity.Spec.Status), "PipelineActivity status for %s", activity.Name)
-	assert.Equal(t, string(v1.ActivityStatusTypeSucceeded), string(activity.Spec.WorkflowStatus), "PipelineActivity workflow status for %s", activity.Name)
-	for _, step := range activity.Spec.Steps {
-		promote := step.Promote
-		if promote != nil {
-			assert.Equal(t, string(v1.ActivityStatusTypeSucceeded), string(promote.Status), "PipelineActivity %s status for Promote to Environment %s", activity.Name, promote.Environment)
-		}
-	}
-}
-
-func assertWorkflowStatus(t *testing.T, activities typev1.PipelineActivityInterface, name string, status v1.ActivityStatusType) {
-	activity, err := activities.Get(name, metav1.GetOptions{})
-	if err != nil {
-		assert.NoError(t, err, "Could not find PipelineActivity %s", name)
-		return
-	}
-	if !assert.Equal(t, string(status), string(activity.Spec.Status), "PipelineActivity status for %s", activity.Name) ||
-		!assert.Equal(t, string(status), string(activity.Spec.WorkflowStatus), "PipelineActivity workflow status for %s", activity.Name) {
-		dumpFailedActivity(activity)
-	}
-}
-
-func dumpFailedActivity(activity *v1.PipelineActivity) {
-	data, err := yaml.Marshal(activity)
-	if err == nil {
-		log.Warnf("YAML: %s\n", string(data))
-	}
-}
-
-func assertHasNoPullRequestForEnv(t *testing.T, activities typev1.PipelineActivityInterface, name string, envName string) {
-	activity, err := activities.Get(name, metav1.GetOptions{})
-	if err != nil {
-		assert.NoError(t, err, "Could not find PipelineActivity %s", name)
-		return
-	}
-	for _, step := range activity.Spec.Steps {
-		promote := step.Promote
-		if promote != nil {
-			if promote.Environment == envName {
-				assert.Fail(t, "Should not have a Promote for Environment %s but has %v", envName, promote)
-				return
-			}
-		}
-	}
-}
-
-func createTestPipelineActivity(jxClient versioned.Interface, ns string, folder string, repo string, branch string, build string, workflow string) (*v1.PipelineActivity, error) {
-	activities := jxClient.JenkinsV1().PipelineActivities(ns)
-	key := &kube.PromoteStepActivityKey{
-		PipelineActivityKey: kube.PipelineActivityKey{
-			Name:     folder + "-" + repo + "-" + branch + "-" + build,
-			Pipeline: folder + "/" + repo + "/" + branch,
-			Build:    build,
-		},
-	}
-	a, _, err := key.GetOrCreate(activities)
-	version := "1.0." + build
-	a.Spec.GitOwner = folder
-	a.Spec.GitRepository = repo
-	a.Spec.GitURL = "https://github.com/" + folder + "/" + repo + ".git"
-	a.Spec.Version = version
-	a.Spec.Workflow = workflow
-	_, err = activities.Update(a)
-	return a, err
-}
-
-func CreateFakePullRequest(repository *gits.FakeRepository, env *v1.Environment, modifyRequirementsFn cmd.ModifyRequirementsFn, branchNameText string, title string, message string, pullRequestInfo *cmd.ReleasePullRequestInfo) (*cmd.ReleasePullRequestInfo, error) {
-	if pullRequestInfo == nil {
-		pullRequestInfo = &cmd.ReleasePullRequestInfo{}
-	}
-
-	if pullRequestInfo.PullRequest == nil {
-		pullRequestInfo.PullRequest = &gits.GitPullRequest{}
-	}
-	pr := pullRequestInfo.PullRequest
-	if pr.Number == nil {
-		repository.PullRequestCounter++
-		n := repository.PullRequestCounter
-		pr.Number = &n
-	}
-	if pr.URL == "" {
-		n := *pr.Number
-		pr.URL = "https://github.com/" + repository.Owner + "/" + repository.Name() + "/pulls/" + strconv.Itoa(n)
-	}
-	if pr.Owner == "" {
-		pr.Owner = repository.Owner
-	}
-	if pr.Repo == "" {
-		pr.Repo = repository.Name()
-	}
-
-	log.Infof("Creating fake Pull Request for env %s branch %s title %s message %s with number %d and URL %s\n", env.Name, branchNameText, title, message, *pr.Number, pr.URL)
-
-	if pr != nil && pr.Number != nil {
-		n := *pr.Number
-		log.Infof("Creating fake PullRequest number %d at URL %s\n", n, pr.URL)
-
-		// lets add a pending commit too
-		commitSha := string(uuid.NewUUID())
-		commit := &gits.FakeCommit{
-			Commit: &gits.GitCommit{
-				SHA:     commitSha,
-				Message: "dummy commit " + commitSha,
-			},
-			Status: gits.CommitStatusPending,
-		}
-
-		repository.PullRequests[n] = &gits.FakePullRequest{
-			PullRequest: pr,
-			Commits:     []*gits.FakeCommit{commit},
-			Comment:     "comment for PR",
-		}
-		repository.Commits = append(repository.Commits, commit)
-	} else {
-		log.Warnf("Missing number for PR %s\n", pr.URL)
-	}
-	return pullRequestInfo, nil
-}
-
-func NewCreateEnvPullRequestFn(provider *gits.FakeProvider) cmd.CreateEnvPullRequestFn {
-	fakePrFn := func(env *v1.Environment, modifyRequirementsFn cmd.ModifyRequirementsFn, branchNameText string, title string, message string, pullRequestInfo *cmd.ReleasePullRequestInfo) (*cmd.ReleasePullRequestInfo, error) {
-		envURL := env.Spec.Source.URL
-		values := []string{}
-		for _, repos := range provider.Repositories {
-			for _, repo := range repos {
-				cloneURL := repo.GitRepo.CloneURL
-				if cloneURL == envURL {
-					return CreateFakePullRequest(repo, env, modifyRequirementsFn, branchNameText, title, message, pullRequestInfo)
-				}
-				values = append(values, cloneURL)
-			}
-		}
-		return nil, fmt.Errorf("Could not find repository for cloneURL %s values found %s", envURL, strings.Join(values, ", "))
-	}
-	return fakePrFn
 }
